Resolve compiler warning

+++ src/masternodes/govvariables/attributes.cpp
@@ -1209,7 +1209,7 @@ Res ATTRIBUTES::Apply(CCustomCSView & mnview, const uint32_t height)
                     }
                 }
             } else if (attrV0->key == TokenKeys::LoanCollateralFactor) {
-                if (height >= Params().GetConsensus().FortCanningEpilogueHeight) {
+                if (height >= static_cast<uint32_t>(Params().GetConsensus().FortCanningEpilogueHeight)) {
                     std::set<CAmount> ratio;
                     mnview.ForEachLoanScheme([&ratio](const std::string &identifier, const CLoanSchemeData &data) {
                         ratio.insert(data.ratio);
